+++ test/components/component_test.rb
@@ -7,6 +7,7 @@ class PrimerComponentTest < Minitest::Test
 
   # Components with any arguments necessary to make them render
   COMPONENTS_WITH_ARGS = [
+    [Primer::Alpha::UnderlinePanels, { label: "Panel label" }],
     [Primer::Image, { src: "src", alt: "alt" }],
     [Primer::LocalTime, { datetime: DateTime.parse("2014-06-01T13:05:07Z") }],
     [Primer::ImageCrop, { src: "Foo" }],
@@ -22,7 +23,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::BlankslateComponent, { title: "Foo" }],
     [Primer::BorderBoxComponent, {}, proc { |component| component.header { "Foo" } }],
     [Primer::BoxComponent, {}],
-    [Primer::BreadcrumbComponent, {}, proc { |component| component.item { "Foo" } }],
+    [Primer::Beta::Breadcrumbs, {}, proc { |component| component.item { "Foo" } }],
     [Primer::ButtonComponent, {}],
     [Primer::ButtonGroup, {}, proc { |component| component.button { "Button" } }],
     [Primer::Alpha::ButtonMarketing, {}],
@@ -66,7 +67,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::TimeAgoComponent, { time: Time.zone.now }],
     [Primer::TimelineItemComponent, {}, proc { |component| component.body { "Foo" } }],
     [Primer::Tooltip, { label: "More" }],
-    [Primer::UnderlineNavComponent, { label: "aria label" }, proc { |component| component.tab(selected: true) { "Foo" } }]
+    [Primer::Alpha::UnderlineNav, { label: "aria label" }, proc { |component| component.tab(selected: true) { "Foo" } }]
   ].freeze
 
   def test_registered_components
@@ -113,34 +114,6 @@ class PrimerComponentTest < Minitest::Test
     COMPONENTS_WITH_ARGS.each do |component, args, proc|
       render_component(component, default_args.merge(args), proc)
       assert_selector("[data-ga-click='Foo,bar']", visible: false)
-    end
-  end
-
-  def test_all_slots_support_system_arguments
-    COMPONENTS_WITH_ARGS.each do |component, args|
-      next unless component.respond_to?(:slots) && component.slots.any?
-
-      result = render_inline(component.new(**args)) do |c|
-        component.slots.each do |slot_name, _slot_attributes|
-          c.slot(
-            slot_name,
-            classes: "test-#{slot_name}",
-            my: 1,
-            hidden: true,
-            style: "height: 100%;",
-            "data-ga-click": "Foo,bar"
-          ) { "foo" }
-        end
-      end
-
-      component.slots.each do |slot_name, _attrs|
-        assert_selector(
-          ".test-#{slot_name}.my-1[hidden][data-ga-click='Foo,bar']",
-          visible: false
-        )
-
-        assert_includes result.to_html, "height: 100%;"
-      end
     end
   end