+++ test/components/navigation/tab_component_test.rb
@@ -40,8 +40,10 @@ class PrimerNavigationTabComponentTest < Minitest::Test
   def test_raises_if_with_panel_and_no_panel_id
     err = assert_raises ArgumentError do
       render_inline Primer::Navigation::TabComponent.new(with_panel: true) do |c|
+        # :nocov:
         c.text { "Title" }
         c.panel { "content" }
+        # :nocov:
       end
     end
 
@@ -137,7 +139,7 @@ class PrimerNavigationTabComponentTest < Minitest::Test
       c.text { "Title" }
     end
 
-    assert_selector("li.d-flex") do
+    assert_selector("li.d-inline-flex") do
       assert_selector("a") do
         assert_selector("span", text: "Title")
       end