+++ test/components/tab_nav_component_test.rb
@@ -8,13 +8,9 @@ class PrimerTabNavComponentTest < Minitest::Test
   def test_raises_if_multiple_tabs_are_selected
     err = assert_raises Primer::TabbedComponentHelper::MultipleSelectedTabsError do
       render_inline(Primer::TabNavComponent.new(label: "label")) do |c|
-        c.tab(selected: true) do
-          "Tab 1"
-        end
+        c.tab(selected: true) { "Tab 1" }
         c.tab { "Tab 2" }
-        c.tab(selected: true) do
-          "Tab 3"
-        end
+        c.tab(selected: true) { "Tab 3" }
       end
     end